requirements.txt
Flask==2.2.3
pyodbc==4.0.32
paramiko==2.10.4
werkzeug==2.2.3
azure-cognitiveservices-vision-face==0.6.1
msrest==0.7.1
Pillow==9.4.0
